# services/speak-gwop/speak-openapi.yaml
openapi: 3.1.0
info:
  title: speak-gwop API
  version: 0.1.0
  description: |
    Standalone agent-native TTS store.

    Flow:
    1) Create Speak identity
    2) Login for Bearer token
    3) Buy credits via Gwop invoice
    4) Claim paid credits
    5) Submit async TTS jobs
    6) Poll job and download artifact URL
servers:
  - url: http://localhost:3020
...
paths:
...
  /v1/credits/invoices:
    post:
      tags: [Credits]
      summary: Create Gwop invoice for a credit SKU
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCreditInvoiceRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCreditInvoiceResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '502':
          $ref: '#/components/responses/Error502'
        '503':
          $ref: '#/components/responses/Error503'
  /v1/orders/{id}:
    get:
      tags: [Credits]
      summary: Fetch local order and related invoice state
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      responses:
        '200':
          description: Order payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrderResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /v1/credits/claim:
    post:
      tags: [Credits]
      summary: Claim credits for a paid invoice
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreditsRequest'
      responses:
        '200':
          description: Credits claimed or already claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimCreditsResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '502':
          $ref: '#/components/responses/Error502'
        '500':
          $ref: '#/components/responses/Error500'
  /v1/voices:
    get:
      tags: [TTS]
      summary: List available voices from active provider
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      responses:
        '200':
          description: Voices listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoicesResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '502':
          $ref: '#/components/responses/Error502'
  /v1/tts/jobs:
    post:
      tags: [TTS]
      summary: Create async TTS job (reserves characters)
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTtsJobRequest'
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTtsJobResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '402':
          $ref: '#/components/responses/Error402'
        '500':
          $ref: '#/components/responses/Error500'
  /v1/tts/jobs/{id}:
    get:
      tags: [TTS]
      summary: Read TTS job status and result
      parameters:
        - $ref: '#/components/parameters/OperatorApiKeyHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
        - SpeakTokenHeader: []
      responses:
        '200':
          description: Job payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TtsJobRecord'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    SpeakTokenHeader:
      type: apiKey
      in: header
      name: x-speak-access-token
  parameters:
    OperatorApiKeyHeader:
      in: header
      name: x-api-key
      required: false
      schema:
        type: string
      description: Required only when deployment sets MASTER_API_KEY.
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error402:
      description: Payment required / insufficient credits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error502:
      description: Upstream/provider failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error503:
      description: Service unavailable / not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    CreateCreditInvoiceRequest:
      type: object
      required: [sku]
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 100
          default: 1
    CreateCreditInvoiceResponse:
      type: object
      required:
        [order_id, speak_agent_id, invoice_id, status, sku, quantity, chars_to_grant, amount_usdc]
      properties:
        order_id:
          type: string
        speak_agent_id:
          type: string
        invoice_id:
          type: string
        status:
          type: string
          enum: [OPEN, PAID, CREDITED, EXPIRED, CANCELED]
        sku:
          type: string
        quantity:
          type: integer
        chars_to_grant:
          type: integer
        amount_usdc:
          type: integer
        pay_url:
          type: string
          nullable: true
        payment_address:
          type: string
          nullable: true
        payment_details_url:
          type: string
        expires_at:
          type: string
          format: date-time
    GetOrderResponse:
      type: object
      required: [order, invoice, warnings]
      properties:
        order:
          type: object
          additionalProperties: true
        invoice:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        warnings:
          type: array
          items:
            type: object
            additionalProperties: true
    ClaimCreditsRequest:
      type: object
      required: [invoice_id]
      properties:
        invoice_id:
          type: string
    ClaimCreditsResponse:
      type: object
      required:
        [speak_agent_id, invoice_id, order_id, order_status, credited_chars, already_credited, characters_remaining]
      properties:
        speak_agent_id:
          type: string
        invoice_id:
          type: string
        order_id:
          type: string
        order_status:
          type: string
          enum: [OPEN, PAID, CREDITED, EXPIRED, CANCELED]
        credited_chars:
          type: integer
        already_credited:
          type: boolean
        characters_remaining:
          type: integer
    VoicesResponse:
      type: object
      required: [provider, voices, count]
      properties:
        provider:
          type: string
        voices:
          type: array
          items:
            type: object
            required: [voice_id, name]
            properties:
              voice_id:
                type: string
              name:
                type: string
              category:
                type: string
              labels:
                type: object
                additionalProperties:
                  type: string
        count:
          type: integer
    CreateTtsJobRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
        voice_id:
          type: string
        model_id:
          type: string
        output_format:
          type: string
        voice_settings:
          type: object
          properties:
            stability:
              type: number
            similarity_boost:
              type: number
            style:
              type: number
            use_speaker_boost:
              type: boolean
    CreateTtsJobResponse:
      type: object
      required: [speak_agent_id, job_id, status, estimated_chars, characters_remaining, poll_url]
      properties:
        speak_agent_id:
          type: string
        job_id:
          type: string
        status:
          type: string
          enum: [queued]
        estimated_chars:
          type: integer
        characters_remaining:
          type: integer
        poll_url:
          type: string
    TtsJobRecord:
      type: object
      required: [id, agent_id, status, created_at, request, usage]
      properties:
        id:
          type: string
        agent_id:
          type: string
        status:
          type: string
          enum: [queued, running, done, failed]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        request:
          type: object
          required: [text, text_length, voice_id, model_id, output_format]
          properties:
            text:
              type: string
            text_length:
              type: integer
            voice_id:
              type: string
            model_id:
              type: string
            output_format:
              type: string
        usage:
          type: object
          required: [estimated_chars, reserved_chars]
          properties:
            estimated_chars:
              type: integer
            reserved_chars:
              type: integer
            actual_chars:
              type: integer
            refunded_chars:
              type: integer
        result:
          type: object
          properties:
            download_url:
              type: string
            mime_type:
              type: string
            size_bytes:
              type: integer
            sha256:
              type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...